+++ sdk/src/relay/api/mod.rs
@@ -5,6 +5,7 @@ mod subscribe;
 mod sync;
 mod try_connect;
 mod unsubscribe;
+mod unsubscribe_all;
 
 pub use self::fetch_events::*;
 pub use self::send_event::*;
@@ -13,3 +14,4 @@ pub use self::subscribe::*;
 pub use self::sync::*;
 pub use self::try_connect::*;
 pub use self::unsubscribe::*;
+pub use self::unsubscribe_all::*;